frontend and backend optimizations
