+++ backend/controllers/user.js
@@ -1,0 +1,63 @@
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+
+const User = require("../models/user");
+
+exports.createUser = (req, res, next) => {
+  bcrypt.hash(req.body.password, 5).then((hash) => {
+    const user = new User({
+      email: req.body.email,
+      password: hash,
+    });
+    user
+      .save()
+      .then((result) => {
+        res.status(201).json({
+          message: "user created",
+          result: result,
+        });
+      })
+      .catch((err) => {
+        res.status(500).json({
+          message: "Invalid authentication credentials!!",
+        });
+      });
+  });
+};
+
+exports.userLogin = (req, res, next) => {
+  let fetchedUser;
+  User.findOne({ email: req.body.email })
+    .then((user) => {
+      if (!user) {
+        return res.status(401).json({
+          message: "email not found",
+        });
+      }
+      fetchedUser = user;
+      return bcrypt.compare(req.body.password, user.password);
+    })
+    .then((result) => {
+      if (!result) {
+        return res.status(401).json({
+          message: "incorrect password",
+        });
+      }
+      const token = jwt.sign(
+        { email: fetchedUser.email, userId: fetchedUser._id },
+        process.env.JWT_KEY,
+        { expiresIn: "1hr" }
+      );
+      res.status(200).json({
+        token: token,
+        expiresIn: 3600,
+        userId: fetchedUser._id,
+      });
+    })
+    .catch((err) => {
+      console.log(err);
+      return res.status(401).json({
+        message: "Invalid authentication credentials!!",
+      });
+    });
+};
