+++ src/app/auth/auth.guard.ts
@@ -17,7 +17,7 @@ export class AuthGuard implements CanActivate {
     | Promise<boolean | import('@angular/router').UrlTree> {
     const isAuth = this.authService.getIsAuth();
     if (!isAuth) {
-      this.router.navigate(['/login']);
+      this.router.navigate(['/auth/login']);
     }
     return isAuth;
   }
