+++ src/app/auth/auth.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { AngularMaterialModule } from '../angular-material.moudle';
+
+import { SignupComponent } from './signup/signup.component';
+import { LoginComponent } from './login/login.component';
+import { AuthRoutingModule } from './auth-routing.module';
+
+@NgModule({
+  declarations: [SignupComponent, LoginComponent],
+  imports: [
+    CommonModule,
+    FormsModule,
+    AngularMaterialModule,
+    AuthRoutingModule,
+  ],
+})
+export class AuthModule {}
