+++ src/app/posts/posts.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
+import { PostCreateComponent } from './post-create/post-create.component';
+import { PostListComponent } from './post-list/post-list.component';
+import { AngularMaterialModule } from '../angular-material.moudle';
+
+@NgModule({
+  declarations: [PostCreateComponent, PostListComponent],
+  imports: [
+    CommonModule,
+    AngularMaterialModule,
+    ReactiveFormsModule,
+    RouterModule,
+  ],
+})
+export class PostsModule {}
